+++ test/Main.hs
@@ -131,6 +131,12 @@ tests = testGroup "Bytes"
       case Bytes.splitTwice 0xEF (slicedPack (xs ++ [0xEF] ++ ys ++ [0xEF] ++ zs)) of
         Just r -> r === (slicedPack xs, slicedPack ys, slicedPack zs)
         Nothing -> property False
+  , testProperty "splitThrice" $ \(ws :: [Word8]) (xs :: [Word8]) (ys :: [Word8]) (zs :: [Word8])->
+      (all (/=0xEF) ws && all (/=0xEF) xs && all (/=0xEF) ys && all (/=0xEF) zs)
+      ==>
+      case Bytes.splitThrice 0xEF (slicedPack (ws ++ [0xEF] ++ xs ++ [0xEF] ++ ys ++ [0xEF] ++ zs)) of
+        Just r -> r === (slicedPack ws, slicedPack xs, slicedPack ys, slicedPack zs)
+        Nothing -> property False
   ]
 
 bytes :: String -> Bytes